# Optimize 2D lights using specialization constants in RD renderer