--- a/servers/rendering/renderer_rd/shaders/canvas.glsl
+++ b/servers/rendering/renderer_rd/shaders/canvas.glsl
@@ -524,8 +524,7 @@ void main() {
 		color *= texture(sampler2D(color_texture, texture_sampler), uv);
 	}
 
-	uint light_count = draw_data.flags & 15u; //max 15 lights
-	bool using_light = (light_count + canvas_data.directional_light_count) > 0;
+	bool using_light = sc_use_directional_lighting() && sc_light_count() > 0;
 
 	vec3 normal;
 
@@ -607,7 +606,7 @@ void main() {
 #endif
 
 #if !defined(MODE_UNSHADED)
-	if (sc_use_lighting()) {
+	if (sc_use_directional_lighting()) {
 		// Directional Lights
 
 		for (uint i = 0; i < canvas_data.directional_light_count; i++) {
@@ -648,95 +647,93 @@ void main() {
 			light_only_alpha += light_color.a;
 #endif
 		}
+	}
 
-		// Positional Lights
+	// Positional Lights
 
-		for (uint i = 0; i < MAX_LIGHTS_PER_ITEM; i++) {
-			if (i >= light_count) {
-				break;
-			}
-			uint light_base = bitfieldExtract(draw_data.lights[i >> 2], (int(i) & 0x3) * 8, 8);
+	for (uint i = 0; i < sc_light_count(); i++) {
+		uint light_base = bitfieldExtract(draw_data.lights[i >> 2], (int(i) & 0x3) * 8, 8);
 
-			vec2 tex_uv = (vec4(vertex, 0.0, 1.0) * mat4(light_array.data[light_base].texture_matrix[0], light_array.data[light_base].texture_matrix[1], vec4(0.0, 0.0, 1.0, 0.0), vec4(0.0, 0.0, 0.0, 1.0))).xy; //multiply inverse given its transposed. Optimizer removes useless operations.
-			vec2 tex_uv_atlas = tex_uv * light_array.data[light_base].atlas_rect.zw + light_array.data[light_base].atlas_rect.xy;
+		vec2 tex_uv = (vec4(vertex, 0.0, 1.0) * mat4(light_array.data[light_base].texture_matrix[0], light_array.data[light_base].texture_matrix[1], vec4(0.0, 0.0, 1.0, 0.0), vec4(0.0, 0.0, 0.0, 1.0))).xy; //multiply inverse given its transposed. Optimizer removes useless operations.
+		vec2 tex_uv_atlas = tex_uv * light_array.data[light_base].atlas_rect.zw + light_array.data[light_base].atlas_rect.xy;
 
-			if (any(lessThan(tex_uv, vec2(0.0, 0.0))) || any(greaterThanEqual(tex_uv, vec2(1.0, 1.0)))) {
-				//if outside the light texture, light color is zero
-				continue;
-			}
+		if (any(lessThan(tex_uv, vec2(0.0, 0.0))) || any(greaterThanEqual(tex_uv, vec2(1.0, 1.0)))) {
+			//if outside the light texture, light color is zero
+			continue;
+		}
 
-			vec4 light_color = textureLod(sampler2D(atlas_texture, texture_sampler), tex_uv_atlas, 0.0);
-			vec4 light_base_color = light_array.data[light_base].color;
+		vec4 light_color = textureLod(sampler2D(atlas_texture, texture_sampler), tex_uv_atlas, 0.0);
+		vec4 light_base_color = light_array.data[light_base].color;
 
 #ifdef LIGHT_CODE_USED
 
-			vec4 shadow_modulate = vec4(1.0);
-			vec3 light_position = vec3(light_array.data[light_base].position, light_array.data[light_base].height);
+		vec4 shadow_modulate = vec4(1.0);
+		vec3 light_position = vec3(light_array.data[light_base].position, light_array.data[light_base].height);
 
-			light_color.rgb *= light_base_color.rgb;
-			light_color = light_compute(light_vertex, light_position, normal, light_color, light_base_color.a, specular_shininess, shadow_modulate, screen_uv, uv, base_color, false);
+		light_color.rgb *= light_base_color.rgb;
+		light_color = light_compute(light_vertex, light_position, normal, light_color, light_base_color.a, specular_shininess, shadow_modulate, screen_uv, uv, base_color, false);
 #else
 
-			light_color.rgb *= light_base_color.rgb * light_base_color.a;
+		light_color.rgb *= light_base_color.rgb * light_base_color.a;
 
-			if (normal_used) {
-				vec3 light_pos = vec3(light_array.data[light_base].position, light_array.data[light_base].height);
-				vec3 pos = light_vertex;
-				vec3 light_vec = normalize(light_pos - pos);
+		if (normal_used) {
+			vec3 light_pos = vec3(light_array.data[light_base].position, light_array.data[light_base].height);
+			vec3 pos = light_vertex;
+			vec3 light_vec = normalize(light_pos - pos);
 
-				light_color.rgb = light_normal_compute(light_vec, normal, base_color.rgb, light_color.rgb, specular_shininess, specular_shininess_used);
-			} else {
-				light_color.rgb *= base_color.rgb;
-			}
-#endif
-
-			if (bool(light_array.data[light_base].flags & LIGHT_FLAGS_HAS_SHADOW) && bool(draw_data.flags & (INSTANCE_FLAGS_SHADOW_MASKED << i))) {
-				vec2 shadow_pos = (vec4(shadow_vertex, 0.0, 1.0) * mat4(light_array.data[light_base].shadow_matrix[0], light_array.data[light_base].shadow_matrix[1], vec4(0.0, 0.0, 1.0, 0.0), vec4(0.0, 0.0, 0.0, 1.0))).xy; //multiply inverse given its transposed. Optimizer removes useless operations.
-
-				vec2 pos_norm = normalize(shadow_pos);
-				vec2 pos_abs = abs(pos_norm);
-				vec2 pos_box = pos_norm / max(pos_abs.x, pos_abs.y);
-				vec2 pos_rot = pos_norm * mat2(vec2(0.7071067811865476, -0.7071067811865476), vec2(0.7071067811865476, 0.7071067811865476)); //is there a faster way to 45 degrees rot?
-				float tex_ofs;
-				float distance;
-				if (pos_rot.y > 0) {
-					if (pos_rot.x > 0) {
-						tex_ofs = pos_box.y * 0.125 + 0.125;
-						distance = shadow_pos.x;
-					} else {
-						tex_ofs = pos_box.x * -0.125 + (0.25 + 0.125);
-						distance = shadow_pos.y;
-					}
-				} else {
-					if (pos_rot.x < 0) {
-						tex_ofs = pos_box.y * -0.125 + (0.5 + 0.125);
-						distance = -shadow_pos.x;
-					} else {
-						tex_ofs = pos_box.x * 0.125 + (0.75 + 0.125);
-						distance = -shadow_pos.y;
-					}
-				}
-
-				distance *= light_array.data[light_base].shadow_zfar_inv;
-
-				//float distance = length(shadow_pos);
-				vec4 shadow_uv = vec4(tex_ofs, light_array.data[light_base].shadow_y_ofs, distance, 1.0);
-
-				light_color = light_shadow_compute(light_base, light_color, shadow_uv
-#ifdef LIGHT_CODE_USED
-						,
-						shadow_modulate.rgb
-#endif
-				);
-			}
-
-			light_blend_compute(light_base, light_color, color.rgb);
-#ifdef MODE_LIGHT_ONLY
-			light_only_alpha += light_color.a;
-#endif
+			light_color.rgb = light_normal_compute(light_vec, normal, base_color.rgb, light_color.rgb, specular_shininess, specular_shininess_used);
+		} else {
+			light_color.rgb *= base_color.rgb;
 		}
-	}
 #endif
+
+		if (bool(light_array.data[light_base].flags & LIGHT_FLAGS_HAS_SHADOW) && bool(draw_data.flags & (INSTANCE_FLAGS_SHADOW_MASKED << i))) {
+			vec2 shadow_pos = (vec4(shadow_vertex, 0.0, 1.0) * mat4(light_array.data[light_base].shadow_matrix[0], light_array.data[light_base].shadow_matrix[1], vec4(0.0, 0.0, 1.0, 0.0), vec4(0.0, 0.0, 0.0, 1.0))).xy; //multiply inverse given its transposed. Optimizer removes useless operations.
+
+			vec2 pos_norm = normalize(shadow_pos);
+			vec2 pos_abs = abs(pos_norm);
+			vec2 pos_box = pos_norm / max(pos_abs.x, pos_abs.y);
+			vec2 pos_rot = pos_norm * mat2(vec2(0.7071067811865476, -0.7071067811865476), vec2(0.7071067811865476, 0.7071067811865476)); //is there a faster way to 45 degrees rot?
+			float tex_ofs;
+			float distance;
+			if (pos_rot.y > 0) {
+				if (pos_rot.x > 0) {
+					tex_ofs = pos_box.y * 0.125 + 0.125;
+					distance = shadow_pos.x;
+				} else {
+					tex_ofs = pos_box.x * -0.125 + (0.25 + 0.125);
+					distance = shadow_pos.y;
+				}
+			} else {
+				if (pos_rot.x < 0) {
+					tex_ofs = pos_box.y * -0.125 + (0.5 + 0.125);
+					distance = -shadow_pos.x;
+				} else {
+					tex_ofs = pos_box.x * 0.125 + (0.75 + 0.125);
+					distance = -shadow_pos.y;
+				}
+			}
+
+			distance *= light_array.data[light_base].shadow_zfar_inv;
+
+			//float distance = length(shadow_pos);
+			vec4 shadow_uv = vec4(tex_ofs, light_array.data[light_base].shadow_y_ofs, distance, 1.0);
+
+			light_color = light_shadow_compute(light_base, light_color, shadow_uv
+#ifdef LIGHT_CODE_USED
+					,
+					shadow_modulate.rgb
+#endif
+			);
+		}
+
+		light_blend_compute(light_base, light_color, color.rgb);
+#ifdef MODE_LIGHT_ONLY
+		light_only_alpha += light_color.a;
+#endif
+	}
+
+#endif // !MODE_UNSHADED
 
 #ifdef MODE_LIGHT_ONLY
 	color.a *= light_only_alpha;