This builds on our introduction of Ubershaders. By making positional light count a specialization constant we allow the shader compiler to optimize the light loop by either unrolling or not. At low values for `light_count` this results in an improvement to performance.

This breaks batching when light count changes between items, which can increase the CPU cost of rendering. Especially when there are many, small lights. However, since the lighting code is heavily GPU ALU bound, this is a net win for games using 2D lighting.

Marking as 4.x for now as I want to get much broader testing before concluding that this is a net win. So far I have only seen equal or better performance. But it would be nice to compare in real projects. And I would like to compare after we merge https://github.com/godotengine/godot/pull/100302

### Performance testing

I tested this little project on 3 devices: 
- M2 MBP
- Pixel 4 Adreno 640 
- Intel Core i7-1165G7 laptop (integrated graphics)
[light2dopt.zip](https://github.com/user-attachments/files/18161086/light2dopt.zip)

I tested 3 different scenarios:
1. Single light touching many sprites
2. Singly light touching many sprites, second light touching some
3. Many lights touching many sprites

In all cases I saw the same or better performance (which shows that all these devices are heavily GPU-bound when using 2D lighting)

I expect the performance benefit will be negligible on devices with powerful dedicated GPUs and most valuable on mobile devices

**M2 MBP**
Single light: 795 FPS -> 1130 FPS
Many lights: 212 FPS -> 212 FPS

**Pixel 4**
Single light: 31 FPS -> 39 FPS
_Note: The single light case is vsync capped on this device, so it was modified to have 4x the number of sprites and 4 lights (not overlapping)_
Many lights: 32 FPS-> 34 FPS
Isometric demo: 15 FPS -> 17 FPS

**Intel iGPU**
Single light: 263 FPS -> 277 FPS
Many lights: 117 FPS -> 120 FPS
Isometric demo: 140 -> 140