@clayjohn 2025-01-02
I agree. I should have made this a draft PR from the start. 

My concerns with this approach are:
1. CPU cost from reduced batching
2. Pipeline stutters from lack of ubershader. 

Neither are a problem in my test scenes, but I suspect they will be issues for real projects. 

This PR is a proof of concept. The performance improvement wasn't as big as I expected, so I would like to explore alternative options